# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Klinik Sehat API",
        "description": "REST API untuk aplikasi manajemen klinik. Mencakup manajemen dokter, jadwal, reservasi pasien, dan rekam medis.",
        "contact": {
            "name": "Klinik Sehat Dev Team",
            "email": "admin@klinik.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/rekam-medis": {
            "get": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Semua rekam medis (admin only)",
                "operationId": "93987d0b674378458458354edbe6b40d",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Cari diagnosis atau resep",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "dokter_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar rekam medis"
                    },
                    "403": {
                        "description": "Bukan admin"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Buat rekam medis baru (admin only)",
                "description": "user_id dan dokter_id diambil otomatis dari reservasi_id. Status reservasi otomatis berubah ke 'selesai'.",
                "operationId": "701eaff472e843aea292b35e8af752d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reservasi_id",
                                    "tanggal_periksa",
                                    "diagnosis",
                                    "resep_obat"
                                ],
                                "properties": {
                                    "reservasi_id": {
                                        "description": "Unik — satu reservasi hanya bisa punya satu rekam medis",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tanggal_periksa": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-06-10"
                                    },
                                    "diagnosis": {
                                        "type": "string",
                                        "example": "Infeksi Saluran Pernapasan Atas (ISPA)",
                                        "minLength": 5
                                    },
                                    "resep_obat": {
                                        "type": "string",
                                        "example": "Paracetamol 500mg 3x1",
                                        "minLength": 5
                                    },
                                    "catatan_dokter": {
                                        "type": "string",
                                        "example": "Istirahat 3 hari."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rekam medis berhasil dibuat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Rekam medis berhasil ditambahkan. Status reservasi diperbarui ke 'selesai'."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Bukan admin"
                    },
                    "422": {
                        "description": "Validasi gagal atau reservasi sudah punya rekam medis"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        },
        "/rekam-medis/{id}": {
            "get": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Detail rekam medis (admin only)",
                "operationId": "cb47d9b49eef69cb5c6934b81f02868c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail rekam medis"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Update rekam medis (admin only)",
                "operationId": "8f184343fecf141f00bb98f02ac4bd9d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "diagnosis": {
                                        "type": "string",
                                        "example": "Diagnosis diperbarui"
                                    },
                                    "resep_obat": {
                                        "type": "string",
                                        "example": "Resep diperbarui"
                                    },
                                    "catatan_dokter": {
                                        "type": "string",
                                        "example": "Catatan diperbarui"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rekam medis berhasil diperbarui"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Rekam Medis"
                ],
                "summary": "Hapus rekam medis (admin only)",
                "operationId": "9dc21d07c096214125d0ae711f7a7798",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rekam medis berhasil dihapus"
                    },
                    "404": {
                        "description": "Tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "apiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan JWT token dari response login. Format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "apiKeyAuth": {
                "type": "apiKey",
                "description": "API Key dari response login/register. Wajib untuk endpoint admin.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}